+++ test/compatibility/v4/lib/resourceStore.js
@@ -1,0 +1,9 @@
+// eslint-disable-next-line import/prefer-default-export
+export function getResourceBundle(lng, ns) {
+  if (!ns) ns = this.options.defaultNS;
+
+  // COMPATIBILITY: remove extend in v2.1.0
+  if (this.options.compatibilityAPI === 'v1') return { ...{}, ...this.getResource(lng, ns) };
+
+  return this.getResource(lng, ns);
+}